[fix]constant 삭제
